redisgned metascope database layer
