--- schedoscope-metascope/src/main/java/org/schedoscope/metascope/model/ActivityEntity.java
@@ -1,70 +1,0 @@
-/**
- * Copyright 2015 Otto (GmbH & Co KG)
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.schedoscope.metascope.model;
-
-import org.schedoscope.metascope.model.key.ActivityEntityKey;
-
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-
-@Entity
-public class ActivityEntity {
-
-    public enum ActivityType {
-        NEW_COMMENT, UPDATE_DOCUMENTATION, UPDATE_TAGS, UPDATE_TABLE_METADATA
-    }
-
-    @EmbeddedId
-    private ActivityEntityKey key;
-    private long timestamp;
-
-    public ActivityEntity() {
-        this.key = new ActivityEntityKey();
-    }
-
-    public ActivityType getType() {
-        return key.getType();
-    }
-
-    public void setType(ActivityType type) {
-        this.key.setType(type);
-    }
-
-    public String getUsername() {
-        return key.getUsername();
-    }
-
-    public void setUsername(String username) {
-        this.key.setUsername(username);
-    }
-
-    public TableEntity getTable() {
-        return key.getTable();
-    }
-
-    public void setTable(TableEntity table) {
-        this.key.setTable(table);
-    }
-
-    public long getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(long timestamp) {
-        this.timestamp = timestamp;
-    }
-
-}
