--- schedoscope-metascope/src/main/java/org/schedoscope/metascope/model/ExportEntity.java
@@ -1,69 +1,0 @@
-/**
- * Copyright 2015 Otto (GmbH & Co KG)
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.schedoscope.metascope.model;
-
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.util.Map;
-
-@Entity
-public class ExportEntity {
-
-    public static final String EXPORT_ID = "export_id";
-    public static final String FQDN = "fqdn";
-    public static final String TYPE = "type";
-
-    @Id
-    private String exportId;
-    private String fqdn;
-    private String type;
-    @ElementCollection
-    private Map<String, String> properties;
-
-    public String getExportId() {
-        return exportId;
-    }
-
-    public void setExportId(String exportId) {
-        this.exportId = exportId;
-    }
-
-    public String getFqdn() {
-        return fqdn;
-    }
-
-    public void setFqdn(String fqdn) {
-        this.fqdn = fqdn;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public Map<String, String> getProperties() {
-        return properties;
-    }
-
-    public void setProperties(Map<String, String> properties) {
-        this.properties = properties;
-    }
-
-}
