--- schedoscope-metascope/src/main/java/org/schedoscope/metascope/model/key/FieldEntityKey.java
@@ -1,89 +1,0 @@
-/**
- * Copyright 2015 Otto (GmbH & Co KG)
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.schedoscope.metascope.model.key;
-
-import org.schedoscope.metascope.model.FieldEntity;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.io.Serializable;
-
-@Embeddable
-public class FieldEntityKey implements Serializable {
-
-    private static final long serialVersionUID = 6389851840855124146L;
-
-    @Column(name = FieldEntity.FQDN)
-    private String fqdn;
-    @Column(name = FieldEntity.NAME)
-    private String name;
-
-    public FieldEntityKey() {
-    }
-
-    public FieldEntityKey(String fqdn, String name) {
-        this.fqdn = fqdn;
-        this.name = name;
-    }
-
-    public String getFqdn() {
-        return fqdn;
-    }
-
-    public void setFqdn(String fqdn) {
-        this.fqdn = fqdn;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((fqdn == null) ? 0 : fqdn.hashCode());
-        result = prime * result + ((name == null) ? 0 : name.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        FieldEntityKey other = (FieldEntityKey) obj;
-        if (fqdn == null) {
-            if (other.fqdn != null)
-                return false;
-        } else if (!fqdn.equals(other.fqdn))
-            return false;
-        if (name == null) {
-            if (other.name != null)
-                return false;
-        } else if (!name.equals(other.name))
-            return false;
-        return true;
-    }
-
-}
